requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
redis==5.0.1
python-multipart==0.0.7
google-generativeai==0.3.2
numpy==1.24.3
python-dotenv==1.0.0
langchain
pydantic_settings